Refactored `suggestion-table` `useEcho` hooks #34 prevents code repeats
